rearrange the composites ui elements routes
