+++ src/core/layout/Uim.tsx
@@ -6,7 +6,7 @@ import Modal from "@mui/material/Modal";
 
 import { Login } from "core/layout/uim/Login";
 import { UserCard } from "core/layout/uim/UserInfo";
-import ConfirmationModal from "components/ConfirmationModal";
+import { ConfirmationModal } from "core/composites/ConfirmationModal";
 import { useUserCTX } from "core/hooks/UserContext";
 import { deleteTokensFromLS } from "core/utils/JWTstorage";
