+++ src/core/utils/externalRequests.js
