+++ src/core/utils/geoServerAPI.js
