+++ src/pages/search/dashboard/Ecosystems.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import TextBoxes from "pages/search/shared_components/TextBoxes";
 import {
