+++ src/pages/search/dashboard/landscape/connectivity/CurrentPAConnectivity.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
 
 import { PointFilledLegend } from "pages/search/shared_components/CssLegends";
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import {
   SearchLegacyCTX,
