+++ src/pages/search/dashboard/landscape/connectivity/CurrentSEPAConnectivity.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
 
 import DownloadCSV from "pages/search/shared_components/DownloadCSV";
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import {
   SearchLegacyCTX,
