+++ src/pages/search/dashboard/landscape/forest/ForestLossPersistence.tsx
@@ -5,7 +5,7 @@ import {
   SearchLegacyCTX,
   type LegacyContextValues,
 } from "pages/search/SearchContext";
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import matchColor from "pages/search/utils/matchColor";
 import formatNumber from "pages/search/utils/format";
