+++ src/pages/search/dashboard/landscape/humanFootprint/CurrentFootprint.tsx
@@ -6,7 +6,7 @@ import {
   SearchLegacyCTX,
   type LegacyContextValues,
 } from "pages/search/SearchContext";
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import matchColor from "pages/search/utils/matchColor";
 import BackendAPI from "pages/search/utils/backendAPI";
