+++ src/pages/search/dashboard/landscape/humanFootprint/TimelineFootprint.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
 
-import ShortInfo from "components/ShortInfo";
+import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import {
   SearchLegacyCTX,
